removed data
Signed-off-by: Kent Madsen <kent.vejrup.madsen@designermadsen.dk>

Initial commit

Initial commit

addition of data

addition of data

adding data

Renaming & adding data

renaming of files to follow convention

adding workflow

working on flows

working on labels

working on labels

removed formats, will automate the process at some point

working on workflow

added data for 31 08 2022

added data for 31 08 2022

01 september 2022

added data for 2022 09 02

s

added data for 2022 09 02

added data for 03 09 2022

added data for 04 - 09 2022

removing spaces from license. not necesarry as it is md and will do it automaticly for me

removing spaces, will be making it readable later

updating readme.md

updating workflow

working on documentation

working on documentation

r

r

working with md

working with documentation

removed space

transition

renamed labels

test

typo

updating documentation

added hashes

editing documentation

doc

moved

added hashes

added footnotes

updating documentation

added last data for 06 09 2022

added data

adding data for 08 and today, had synch. issues with the clock today because of dual boot

adding data for 09, 12 and the 13th september 2022

values normalised

added data from 14 sep. 2022  from openhardware

added data for 15 09 2022

renamed so it only represents the date it was recorded

added data for today

renamed file to the day of the month it was recorded and added data for 17 09 2022

added data for today

renamed to date in month

added data for yesterday

appended

added data for 19 09 2022

renamed to the day in the month of recording

added data for 20 09 2022

renamed to date of recording

added data for 21 09 2022

renamed to date of recording

adding data for 2022 09 22

renamed to date of recording

data added

added data for 24 & 23 09 2022

renamed to appropiate date

added data for 25 09 2022

renamed to appropiate date

added data for 26 - 09 - 2022

renamed to date

27 - 28 09 2022 added data

renamed to date

added data for 2022 09 29

renamed to date

c

30 09 2022 added

renamed to the date

01-10-2022

added data for 02-10-2022

added data for 03 - 10 - 2022

renamed to date of recording

added data for 04 10 2022

added data for 05 10 2022

renamed to date of recording

added data for 06 - 10 - 2022

renamed to the date of month, of the recording

added data for 07 10 2022

renamed to date of recording

08 - 10 - 2022

renamed to date of recording

09, 10, 11 - 10 - 2022

renamed to date of recording

added data for 12 - 10 - 2022

renamed to date of recording

added data for 13-10-2022

renamed to date of recording

typo

fixed typo

added data for 14-10-2022 and 15-10-2022

renamed to date of recording

added data for 16 10 2022

renamed to date of recording

added data for 17 and 18 october 2022

renamed to date of recording'

c

added data for 19 10 2022 and renamed to (date in month)

added data recorded at 20 to 22-10-2022

renamed to day of recording

added cover images from https://www.flickr.com/photos/190441416@N05/52405851547/in/dateposted-public/

track large files

updated documents

preparing data

organising archived.

working on workflow

updating changelog

c

added symbolic links